+++ CHANGELOG.md
@@ -3,10 +3,14 @@
 
 ## Upcoming...
 
+## v0.5.0 - 2021-04-22
+
 Breaking changes:
 - StoreToContext middleware split to: StoreUserToContext and StoreParamsToContext
 - Session.SetUserLocale renamed to Session.SetLocale
 - Session.GetUserLocale renamed to Session.GetLocale
+
+Better UI on example app
 
 ## v0.4.0 - 2021-04-20
 
@@ -15,8 +19,9 @@ Breaking changes:
 
 ## v0.3.0 - 2021-04-19
 
+Breaking changes:
+- StoreUserToContext middleware renamed to StoreToContext
+
 - NewMuxRouter constructor added for the underlying mux.Router
 - ContextWithParams / GetParam added for easier input data retrieval
 
-Breaking changes:
-- StoreUserToContext middleware renamed to StoreToContext